+++ notebooks/index.clj
@@ -1014,3 +1014,12 @@ nested-structure-1
              :y2 {:field "petal_length", :type "quantitative"},
              :color {:field "species", :type "nominal"}}
   :background "floralwhite"})
+
+;; ## Hiding code
+
+;; By default, a Clay notebook shows both the code and the result of an evaluated form.
+;; Here are a few ways one may hide the code:
+;;
+;; 1. Add the metadata `:kindly/hide-code true` to the form (e.g., by preceding it with `^:kindly/hide-code`).
+;; 2. Add the metadata `:kindly/hide-code true` to the value.
+;; 3. Globally define certain kinds (e.g., `:kind/md`, `:kind/hiccup`) to always hide code (on project level or namespace level) by adding thme as a set to the project config or namespace config, e.g., `:kindly/options {:kinds-that-hide-code #{:kind/md :kind/hiccup}}`.